requirements.txt
tweepy==3.10.0
requests==2.25.1
beautifulsoup4==4.9.3
Pillow==8.2.0
python-dotenv==0.18.0
